+++ database/migrations/2023_12_11_132699_create_account_request_metas_table.php
@@ -13,27 +13,30 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('account_request_metas', function (Blueprint $table) {
-            $table->id();
+        if(config('filament-accounts.features.requests')){
+            Schema::create('account_request_metas', function (Blueprint $table) {
+                $table->id();
 
-            $table->foreignId('user_id')->nullable()->references('id')->on('users')->onDelete('cascade');
+                $table->foreignId('user_id')->nullable()->references('id')->on('users')->onDelete('cascade');
 
-            $table->unsignedBigInteger('model_id')->nullable();
-            $table->string('model_type')->nullable();
+                $table->unsignedBigInteger('model_id')->nullable();
+                $table->string('model_type')->nullable();
 
-            $table->foreignId('account_request_id')->references('id')->on('account_requests')->onDelete('cascade');
-            $table->string('key')->index();
-            $table->json('value')->nullable();
+                $table->foreignId('account_request_id')->references('id')->on('account_requests')->onDelete('cascade');
+                $table->string('key')->index();
+                $table->json('value')->nullable();
 
-            $table->boolean('is_approved')->default(0)->nullable();
-            $table->datetime('is_approved_at')->nullable();
+                $table->boolean('is_approved')->default(0)->nullable();
+                $table->datetime('is_approved_at')->nullable();
 
-            $table->boolean('is_rejected')->default(0)->nullable();
-            $table->datetime('is_rejected_at')->nullable();
-            $table->text('rejected_reason')->nullable();
+                $table->boolean('is_rejected')->default(0)->nullable();
+                $table->datetime('is_rejected_at')->nullable();
+                $table->text('rejected_reason')->nullable();
 
-            $table->timestamps();
-        });
+                $table->timestamps();
+            });
+        }
+
     }
 
     /**
@@ -43,6 +46,8 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('account_request_metas');
+        if(config('filament-accounts.features.requests')) {
+            Schema::dropIfExists('account_request_metas');
+        }
     }
 };
